requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
azure-keyvault-secrets==4.8.0
azure-identity==1.17.1
Flask-SocketIO=5.3.6
